fix flake8, and yapf
put data_table.py to it's original

passing all the tests

foo

move join_fbi to data directory
